Refactor Cannon.js by adding Tower.js .
